Dedicated namespace for CB v1 scheme
